pyproject.toml
[project]
name = "rubiks-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.0",
    "kociemba>=1.2.1",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pyqt5>=5.15.11",
    "rubik-cube>=0.0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
